Tempus: Improve Observer Names.
Made changes to reduce confusion related to
IntegratorObserver::acceptTimeStep().  acceptTimeStep() checked if
the timestep should be accepted or failed and promoting if passing.
“acceptTimeStep” was being read as a statement, "accept the timestep",
when it was meant as a question, “accept the timestep?”.  To reduce
this confusion, acceptTimeStep has been split into the determination
of the time step passing (now checkTimeStep()) and promotion of the
time step (now promoteWorkingState() is part of the Integrator time
loop).  Also two observers, observeAfterCheckTimeStep() and
observeEndTimeStep, have replaced the observeAcceptedTimeStep().
Work previously completed in observeAcceptedTimeStep() will likely
move into observeEndTimeStep(), unless it deals with the acceptance
of a time step as passing which would go into observeAfterCheckTimeStep().

Passed all tests.