Fix installed load_matching_env.sh for 'chama' and 'serrano' (#3816)
I moved atdm/toss3/ to atdm/common/toss3/ and added install for the
atdm/common/ subdir on all systems.  That fixes the usage of the
load_matching_env.sh script on 'chama' and 'serrano' and provides what I think
is a fairly maintainable solution that still will not allow a user to try to
load some arbitrary env.

I tested the install and sourcing the installed load_matching_env.sh manually
on 'chama' and it seems to do the trick.  I also tested the loading of the
modules on 'serrano' but did not test the install and sourcing
load_matching_env.sh there.  (But since the change for
atdm/serrano/environment.sh was trivial it should work.)