Framework: Splitting PR driver scripts
This splits the PR test driver script into two scripts:
1. PullRequestLinuxDriver.sh
  - This script is the first that should be called by auto-testing
    jobs.  It handles merging the source branch into the target
    branch.
  - The `version` of this script that is run will the one that
    lives on the _target_ branch.
2. PullRequestLinuxDriver-Test.sh
  - This script is called by (1) and handles actually running the
    test.
  - The `version` of this script that is run will be the result
    of the _source_ branch merged into the _target_ branch.

This reduces the _chicken and egg_ problem when updating the testing
system.  Specifically, in the past updating the
`PullRequestLinuxDriver.sh` script gave a false sense of security when
the autotester completed since it only ran the script as it is on the
_target_ branch, without pulling in modifications.  By splitting the
script we limit this issue to just modifications to the _merge_
operation, which is in theory modified much more infrequently than
the test part, which might be adding, modifying, or removing tests
from time to time for dev->master or branch->dev testing.

Resolves #3625