Refactor and reintegrate Hybrid Gmres solver in Belos
The Hybrid (a.k.a. Seed or Poly) Gmres solver has been refactored for ease of
use and extensibility.  There are two polynomials that this solver can generate,
the Arnoldi polynomial and the Gmres polynomial (per Ron Morgan's diction).
Furthermore, it has been integrated into the new Belos solver factory framework
and it also uses the solver factory to generate an outer solver, if requested by
the user.  That required a small correction to the solver factory framework for
complex valued solver factories.

Tests have been added for the Epetra, Tpetra, and basic linear algebra interfaces
to ensure that this solver is tested nightly and works.