Add support for ctest-s-local-test-driver.sh (TRIL-212)
This script makes it super easy to drive builds that submit to CDash on any
machine using node allocations and all of the machinary that drives the ctest
-S builds to submit to CDash.

To support this, I changed ATDM_CONFIG_ALL_SUPPORTED_BUILDS to a bash array
(but the old style works as well) and I added the var
ATDM_CONFIG_CTEST_S_BUILD_NAME_PREFIX to the
<system_name>/all_supported_bulds.sh file.  Now the build names in
ATDM_CONFIG_ALL_SUPPORTED_BUILDS have to exactly match the Jenkins driver bash
scripts.

I updated the all_supported_builds.sh files for all currently supported system
to exactly match the names of the Jenkins driver scripts.