Add sems-rhel7 env with clang, gnu, intel, and cuda compilers (TRIL-236)
The sems-rhel7 env should be selected automatically on any SNL CEO RHEL7
machine that has the SEMS modules mounted.

Currently, the intel-17.0.1 compilers are super slow so this build is
basically unusable (at least on cortado).

There is also a cuda 9.2 + GCC 7.2.0 build that works with
KOKKOS_ARCH=Pascale60 on the cafe machines.

As part of this, I changed the approach a little to improve the error messages
some for what complilers are supported for the sems-rhel7 env.  I also
improved some of the documentation in the README.md.

NOTE: This does not yet add any ctest -S driver scripts due to a problem with
using 'bash -l' on the machine 'cortado' where this env was set up.