Consitently use '.so' for shared and '.a' for static boost shared libs (#3890)
I only tested this on the sems-rhel6 env.  Hopefully this will not cause
problems on other envs.  But note that the only change for other envs is to
use libboost_program_options.a and libboost_system.a when doing the default
'static' builds on the other systems.  We will just cross out fingers that
these libs work correctly on these other systems.