Reverting changes to LAPACK GEES routines to address bug #3992 and #3989
Failures on ATDM platforms started after PR #3951, where the interaction
with the BKS and Generalized Davidson solver and LAPACK GEES changed to
address issues with MKL 18.0.[1-3].  While it seems that checking for the
optimal workspace size should be safe before performing eigensolve in GEES,
it appears there are memory issues with GEES in MKL 17.0.x, where the use
of the optimal memory size is resulting in bad behavior and potential seg
faults.  So, it's a catch 22 and I am going back to the previous version
of these solvers that have been solid for years.