Allow the disable of just the LAPACK test for STEQR() (#2410)
This is the only function that continues to segfault on Power8 and Power9
builds using NETLIB LAPACK in debug builds.  All of the Trilinos code
downstream has removed a dependence on this LAPACK function. This will allow
the rest of the LAPACK tests to be run in Power8 and Power9 'debug' builds.