Change from '-' to '_' to seprate build options in tweak file name (TRIL-212)
This is the form that SPARC likes and it does seem to make some sense.

I also changed from 'RELEASE_DEBUG' to 'RELEASE-DEBUG' to be consistent with
this convention.

I also changed from CUDA-<???>-GNU-<???> to CUDA-<???>_GNU-<???> for the other
builds to match the preference by SPARC.  (This required modifying
ride/environment.sh.)