MueLu: modifying structured aggregation to compute graph of prolongator
The modifies the prolongatorGraph output from structured aggregation.
Instead of getting a node based graph it now returns a dofs based graph.
This means that in the new format, the prolongator graph can be used to call the constructor of the prolongator matrix directly.
This has the advantage of reducing some overhead and it also simplifies the implementation of the geometric interpolation factory.