Revert "Merge pull request #4228 from trilinos/Fix-KokkosKernels-BLAS-mangling"
This reverts commit 24c2eb2f9c622f856a0477a0ef8751b4145ca7b1, reversing
changes made to 7b60b19b90f7cc1066da428b7d8ad18813675a87.

It turns out that this breaks builds if Fortran is not enabled.  CMake
apparently permits the configuration to proceed, but the build fails
with a missing header file.  Here are some issues that were filed as a
result:

https://github.com/kokkos/kokkos-kernels/issues/372
https://github.com/trilinos/Trilinos/issues/4237
https://github.com/trilinos/Trilinos/issues/4235