ShyLU/FROSch: add MueLu tests for Laplace problem
Use MueLu as coarse solver in a TwoLevel_Preconditioner for the Laplace
problems. Enabling these tests depends on enabling MueLu to be used
in the FROSch build configuration.

Note: The MueLu parameter list is somehow artificial as it uses a way
too small "coarse: max size", just to trigger a 2-level MueLu
hierarchy.