Tpetra: add function replaceDiagonalCrsMatrix() Part of #4202, #4297
Implement replacement of diagonal elements as a free function,
that just relies on Tpetra::CrsMatrix's public interface.
The function returns the number of locally replaced diagonal entries.
Check isFillComplete status and existense of column map.
Do map comparison in debug mode only.
A test is added to check correctness in serial and parallel.

Note: this is not thread parallel, yet.