Sacado:  Check for zero base in pow(Fad,Const)
Most Sacado overloads of pow() check for a zero base, and return zero
instead of evaluating the derivative even though in many situations this
isn't technically correct (since the derivative is often not defined
there).  In PR #4380, I changed this behavior for the pow(Fad,Const)
case, where I use a different formula.  This was causing NaN's in
Charon's Jacobian evaluation, so this commit restores the previous
behavior (which again is technically not correct).