Sacdo:  Fix std::pow(Fad,Const) again.
My change of the formula for differentiating std:pow(x,y) with constant
y seems to be causing some numerical stability issues (which is
surprising because it just replaces x^y/x with x^{y-1}).  So I am
reverting that change and going back to the old formula.  I still want
to understand why it is problematic, but it is impacting a customer, so
I am changing it now anyway.  See here for the relevant discussion:

https://cee-gitlab.sandia.gov/Charon/tcad-charon/issues/96#note_179767