MueLu: seperating capabilities in composite_to_regions driver
The idea is to seperate some capabilites from the actual driver in order to develop more generic HHG headers.
These headers can then be included in various driver instead of the current in-place use.
This new header defines hierarchy setup functions and the vCycle.
It could be renamed RegionMG but let's think about that later.
It exposes a better split among various functions that are used in the Region MG algorithm.
A few const qualifiers have been modified to be more accurate.