MueLu/HHG:
1) issue warning for missing coarse solver
Coarse level solver has not been migrated to Xpetra yet.
For now, we just skip it and don't do anything on the coarse level.
However, proc 0 writes a warning message to the screen.

2) Fix map when reading input matrix
Use the composite map instead of creating a linear map when reading
the input matrix.
Remove obsolete function argument

3) Specify example via command line:
The HHG driver reads various input files. The exact example has to be
known in order to setup the structured region MG. Now, the exaxt
name of the example has to be passed via a command line arguement
"--caseName=..."

Add include guards
Move utilities to utils file

Part of #4084.