Remove build cee-rhel6_gnu-4.9.3_openmpi-1.10.2_serial_shared_opt from nighlty testing (TRIL-243)
SPARC does not need this build and since the gnu-7.2.0 build's test suite was
made to run fast (see TRIL-269), we don't need to compare against the old
gnu-4.9.3 build anymore.

Note that I did not remove the env for the GNU 4.9.3 build in case someone
wants to test it locally.