Move load_matching_env.sh into subdir (ATDM)
This avoids having it sit in the source dir along side the load-env.sh script
so tab completion works better and it moves it out of the base build directory
to avoid confusion (because it is only for the install dir).

I tested this manually and it seems to do exactly what it is supposed to do.