Promote cee-rhel6 intel-18.0.2 build with now much faster test suite (#4251, #4260, #4262)
With the changes to the MPI and OpenMP settings suggeted by Brad King in the
last commit, the full test suite now completes in under 9 minutes where before
it was taking 2.5 hours!

Since the test described in #4251, #4260, #4262 don't seem to be timing out
anymore after this change, we can promote this build.

NOTE: The following tests FAILED:

    1477 - STKBalance_stk_balance_MPI_4 (Failed)
    1478 - STKBalance_stk_balance_m2n_MPI_4 (Failed)

But these tests also fail in the promoted cee-rhel6-intel-17.0.1 build so
there is no reason to block the promotion of this build.