Tpetra: Add functions for getting View to statically alloc'd memory
@trilinos/tpetra

Add functions in the Tpetra::Details namespace, that get a View of
statically allocated memory.  This is related to #4626, and is part of
the work-around for the Intel compiler issues observed in PR #4648.

I also added tests that build and pass on CUDA.  The tests include a
work-around for a Kokkos issue:
https://github.com/kokkos/kokkos/issues/2051.