Piro: putting in logic such that when running with the 'Albany Time Step Control Options' ParameterList, Tempus solve is killed upon reaching dt = Mininum Time Step, and encountering a failed solve.
Without this option, the last solve will be repeated (at the min dt, since dt can't be cut further)
nConsecutiveFailuresMax-1 times, which is a waste.

This logic is going in Piro right now and not Tempus because there are corner cases in Tempus for certain
methods (e.g., multi-state, multi-rate methods) that complicates adding the logic to Tempus in a general way.