Disable tests in cuda rdc PT build already failing in non-rdc build (#4502)
This disables tests already failing in the matching non-RDC build so those
failures have nothing to do with RDC.  This will allow us to promote this
build to the "ATDM" group and maintain it.  This build is too expensive to run
as a Trilinos PR build so the next best thing is to run it as a nightly build.

* Disables ROL tests from #3543
* Disables TrilinosCouplings tests from #3749
* Disables Zoltan tests from #4042