Tempus: Improve Interval Output Setup
In the output interval and the screen output interval, all the
intervals are placed in a vector along with the specified output
times or indices, and then sorted. This allowed all the output times
and indices to be in one vector and searched. However when the
number of time steps reaches millions, the sort is far too expensive.
This reworks the interval specification so that it is not part
of the vector or sorting.

 #4836