Set up for alternate workspace, install as atdm-devops, etc. (#2689, ATDV-151, ATDV-156)
This is mainly to allow the usage of installation as the 'atdm-devops-admin'
user account using the 'jenkins' entity account on various systems and
therefore protect the installs of Trilinos from bad 'jenkins' jobs.  There is
a lot little things you need to do to get this work:

* ATDM_CONFIG_WORKSPACE_BASE[_DEFAULT]: Use a different workspace for
  SRC_AND_BUILD (which will allow for reading by the 'atdm-devops-admin'
  account).  (And you must 'cd' into that workspace.)

* ATDM_CONFIG_INSTALL_PBP_RUNNER[_DEFAULT]: Allow inserting the
  'run-as-atdm-devops-admin' setuid program to run the install command as the
  'atdm-devops-admin' user.

* ATDM_CONFIG_TRIL_CMAKE_INSTALL_PREFIX_DATE_BASE[_DEFAULT]: Set a default
  base directory for a given system.

* ATDM_CONFIG_USE_JENKINS_INSTALL_DEFAULTS: Set to '1' to use the defaults for
  the above three vars.

See updated documentation for more details.