IOSS: Update db_zone property after reopening cgns database
The CGNS library will reorder (sort) zones based on their names,
so if an application defines zones to an output file; closes that
file and then reopens the file, the zones on the file will have
a different ordering or indexing than when they were originally
defined on the file.

This commit causes the IOSS library to update the `db_zone` property
on all StructuredBlocks and ElementBlocks when a CGNS database
is closed and then reopened.  It also verifies that all expected
zones (blocks) exist on the database.

Discovered by SPARC team. Verified that fix works for the issue
that they reported. Developing a test to make this continues to work.