MueLu: fix bug in unit-test CreatePreconditioner for #5074
In the epetra adapter: MueLu_CreateEpetraPreconditioner, signature and implementation are split between header and cpp files. This is not the case in the Tpetra path which has only a header.
This implementation detail means that we need to explicitely include <Teuchos_XMLParameterListHelpers.hpp> in CreatePreconditioner.cpp for Epetra only builds.