Tpetra: Harness: Fix withLocalAccess for multiple arguments
@trilinos/tpetra

It had worked before for multiple arguments, when I was using C++14
generic lambdas.  My fix to make withLocalAccess work for C++11 broke
the multiple arguments case.  This commit fixes that.

I found this issue when I wrote an example (actually a test, but it's
really meant to be read) of how to extend withLocalAccess to support
object types other than Tpetra::MultiVector and Tpetra::Vector.  This
commit also includes the example.