Remove mandatory dependence on ParMETIS from stk-balance.
It should be an optional dependence, but it should be an optional
runtime dependence. Since we get the ParMETIS dependency transitively
through Zoltan2, we don't need to mention it at all in stk-balance
cmake files, but rather selectively enable/disable tests that need it.

Also, when a user runs the stk_balance executable it should produce
a semi-graceful message if they try to use a feature that needs
ParMETIS when ParMETIS hasn't been built into Zoltan2.