Xpetra: don't write maps unnecessarily
When writing matrices to disk, the associated maps are written also.
This commit adds a check to see if the domain and range maps are
the same as the row map.  The row map is written unconditionally,
the domain/range maps only if they differ from the row map.
The column map is written only if it differs from the domain map.