MueLu: fixes to region driver to handle multiple dofs per node
Also adding a small change to where numDofsPerNode is set and how dofMap is constructed.
In order to perform computation on Elasticity problems we need to get a nullspace from galeri which prompts the changes described below.
The new class inherits from Problem and is a parent to the other classes defined in Galeri_StencilProblems.hpp. It provides an implementation of BuildNullspace() for the children classes