Change name of env from 'spack-rhel6' to 'spack-rhel'
We know this works with both RHEL6 and RHEL7 with no modifications so no need
to have different envs for RHEL6 and RHEL7.  In fact, this configuration will
likely work for almost any Linux system where the SNL ATDM Spack modules are
installed (but we don't want to claim that just yet without a lot more
testing).