Only print warning about not MUMPS support if MUMPS support is requested (#4915)
ATDM only needs long long global ordinal with Tpetra and is not using MUMPS
with Amesos2.  And yet, Amesos2 was issuing a CMake message(WARNING ...) about
MUMPS not being supported when Amesos2_ENABLE_MUMPS=ON was not even set.

This commit changes things so that warning is only printed if
Amesos2_ENABLE_MUMPS=ON and Tpetra_INST_INT_INT=OFF.

I also combined two if() statements into one with an if(... AND ...) since
CMake supports that just fine.

This is commit was created to remove a bogus warning as part of #4915 that
sets Tpetra_INST_INT_INT=OFF for future ATDM Trilinos builds.