Tpetra::DistObject: Change buffer_memory_space back to CudaSpace
@trilinos/tpetra

The point of changing buffer_memory_space from CudaSpace to
CudaHostPinnedSpace in my previous commit was to make sure that it
builds.  (I had to change a lot of code to make it build.)  I don't
want to leave it that way yet, since we need to test the performance
implications.