MueLu: adding logic in region driver for structured/unstructured runs
Primarily added an new xml file to drive the algorithm correctly.
Also cleaned up a bunch of aggregation related things that were not kept up to date with changes in structured aggregation.
Will probably go at it again especially to clean the constructor of UncoupledIndexManager.
Fixing tests that the changes to HybridAggregation beak.