MueLu: adding SetupRegionSmoother.hpp to region multigrid
This header will provide a unique interface to all smoothers as well as some infrastructure to perform setup and apply phase of all smoothers.
The Jacobi smoother is now moved in the new infrastructure and the run time goes down as the inverse diagonal is cached during setup instead of being recomputed at run time.
Adding the "smootherType" option to the structured region driver to choose one of the available preconditioners.
Reworking the output a little to allow for higher precision and scientific format. Also added a parameter to store residual history to file in case a comparison is wanted.