ATDM: Change from ctest -j8 to -j1 on mutrino
The current settings for srun being used only allow for one srun command to
run a job at a time anyway so might as well avoid bogus timeouts and failures
due to srun timing out before it even starts a job.

The efforts to allow running in parllel on mutrino on the same node have not
worked out as of yet.  See:

    https://gitlab.kitware.com/snl/project-1/issues/91