MueLu: adding an option for reproducible random vector generation in region driver
The idea is to generate the exact same random vector independently of the number of ranks such that the driver run in serial can be compared to the driver run in parallel.

Also reducing the parallel level of nighthly tests on trappist as a few tests are randomly timing out in parallel.