Tpetra::RowMatrixTransposer: Simplify implementation
@trilinos/tpetra

This is part of the fix for #4607.  I eliminated the second, likely
unused code path from createTransposeLocal (that uses the RowMatrix
interface and sequential code to compute the transpose).  That case is
still covered, by conversion of the input RowMatrix to a CrsMatrix.

I also optimized RowMatrixTransposer by adding a constructor to CrsGraph
and CrsMatrix.  The constructor takes the local graph / matrix, all
four Maps, and the Import and Export objects.  It provides the effect of
expertStaticFillComplete, in an idiomatic interface (where construction
is initialization and the object leaves the constructor in a ready
state).