Go back to loading python setup by hand.
The previous shell scripts had unloaded
the seme-python module then added path and
pythonpath directly as the sierra python
modules are not working for the trilinos
entity. I was attempting to fix that as it
would be simpler, but it looks like that is
a no-go on the cloud nodes at this time.