Tpetra:  added KDDHACK2 for further debugging.
Case A:  Do not #define JHUHACK and do not #define KDDHACK2.

This is the original version of the code.  It follows the "if branch"
and calls the fence.  (Output says "KDD If branch" and "KDD Yo fence.")

This case fails as before in FixedHashTable.

Case B:  #define JHUHACK and do not #define KDDHACK2.

This version of the code compiles only the "else" code (without any
branching); the "if branch" and condition test are #ifdef'ed out.
(Output says, "KDD Else branch.")

This cases does NOT fail in a reasonable amount of time.

Case C:  #define KDDHACK2 and do not #define JHUHACK

This version of the code compiles both the "if branch" and the "else
branch," and forces the conditional test to be false so that the "else
branch" is always taken.  (Output says, "KDD Else branch.")

This case fails in with the FixedHashTable error, even though it does
not use the "if branch."

It appears the only difference between Case C and Case B is that Case C
compiles the "if branch" and jumps to the "else branch," whereas Case B
avoids the branching and just executes the "else" code.