Fix smart-jenkins-driver.sh (#4933)
In the previous refactoring to creae get_system_info.sh for #4933, the
variable ATDM_CONFIG_SCRIPT_DIR was newly required to be set but was not set
from a fresh env.

I had tested this smart-jenkins-driver.sh locally but the env var
ATDM_CONFIG_SCRIPT_DIR must have been set in my local env which allowed it to
pass.

This should fix all of the ATDM Trilinos builds.