Testing: rocketman: attempt to speed up nightly tests
Move option Trilinos_CTEST_DO_ALL_AT_ONCE to individual tests.
According to Ross B., setting this to true makes the compile
cycle faster. The trade-off is that if an upstream dependency
fails to compile, the entire test compile cycle will fail (i.e.,
no reconfigure with failing package disabled).
If false, the comple cycle is longer but more robust.  If an
upstream package fails to compile, it's disabled and Trilinos
is reconfigured/recompiled.