Tpetra: fix Multiply without dynamic profile
Fix TpetraExt::MatrixMatrix::Multiply() when deprecated code off.
This function takes C by ref but then overwrites it with A*B.
Since we can't rely on C having been created with the right number of
nonzeros per row, just use the existing mult_A_B_reuse()
and mult_A_B_newmatrix() if the C passed in
is or isn't (respectively) already fillComplete. The newmatrix routine
determines the exact entries per row in the product and handles all the
allocation.

This fixes several test failures in MueLu when Tpetra deprecated is off.