Remove obsolute cee-rhel6 intel 17 build, fix default 'intel' match (SPAR-607)
SPARC is not testing against Intel 17 anymore.  Sierra switched to Intel 19
and SPARC has switched to Intel 19 as well.

I also had to fix the order of matching for "intel"* to be the last intel
elseif or you could never select intel-18 builds because "intel"* would match
"intel-18" but instead would select the intel-19 compiler!