Sacado:  Add some SFINAE to restrict instantiation of ConditionalReturnType
Because of the way the conditional overloads (==, !=, ...) are written
using ConditionalReturnType, this trait can be instantiated with invalid
arguments when the compiler is going through its overload resolution
process for completely unrelated types.  This was observed in SPARC when
using the new Fad design.  This adds some SFINAE metaprogramming to
prevent those instantiations.  It also adds an essentially compile-time
test to make sure the overloads are dropped correctly using SFINAE.