Reduced from ctest -j8 to -j4 for all CUDA builds (#6052)
As described in #6052, it is possible that running with ctest -j8 could cause
out-of-memory errors on the GPU and when using UVM the errors are delayed and
hard to diagnose.  Therefore, to be safe, we need to drop to ctest -j4.  That
will cause all of the 4-proc MPI tests to run by themselves and will allow
four 1-proc MPI tests to run.  Hopefully this will eliminate any possibility
of running out of mememory on the GPU.

Hopefully Kitware can help us create a system where we can more effectively
use the GPUs as part of #2422.