Add new clang-5.0.1-openmpi-1.10.2 stack (CDOFA-44)
For some reason, the BinUtils library libiberty.a gets installed under 'lib'
instead of under 'lib64' when built with clang@5.0.1 vs. gcc@7.2.0 with Spack.

Also, had to pass in absolute path to libgfortran.so to get links with fortran
exectuables to work.  Don't know why that is.  To support this, I added the
env var ATDM_CONFIG_EXTRA_LINK_FLAGS to set Trilinos_EXTRA_LINK_FLAGS.