Muelu drop tolerance:
Machine learning question: Could a per row threshold reduce the number
of solve iterations?

If threshold < 10.0, use normal threshold code path (the threshold only
makes sense in the range [0.0, 1.0].

Otherwise, if 10.0 <= threshold < 100.0, use a random threshold with
(10.0*threshold*row) as the seed to the random number generator.

Otherwise, sort magnitudes into vector A and drop any where
either A[i-1] is dropped or A[i-1]/A[i] > (threshold/100.0).