Tpetra: set packed, 1D, global colinds in CrsGraph
setAllIndices() allows setting the local graph (rowptrs and colinds)
directly. This commit adds setAllGlobalIndices (also an "expert" function),
which does the same thing but takes globally-indexed columns instead.

CrsGraph can now handle globally indexed, STORAGE_1D_PACKED entries.
As with setAllIndices(), k_numRowEntries_ isn't needed.
ConvertColumnIndicesFromGlobalToLocal can now handle the PACKED case.

setAllGlobalIndices() was created to allow the creation of a
fill-complete CrsGraph from the local (but globally indexed) graph,
and the row, domain and range maps. A correct column map is created
automatically by CrsMatrix::fillComplete().
This new function is called makeCrsGraphFromGlobalIndicesAndMaps().

MatrixMatrix::add() now uses this function to create the graph for the
result. If the user didn't request a fill complete result, resumeFill()
is called. This has no downsides because C must be new on input (no space
allocated) and it is allocated to the exact size per row to hold the
sum. The user can freely update scalar values in this state, but not
the sparsity pattern.