Intrepid2: tweaks to OrientationTools::modifyBasisByOrientation() to allow reference-space inputs.
Modified tests for hexahedron, quadrilateral, tetrahedron, and triangle so that these now exercise the new code path as well as the old.  Also tweaked those tests slightly to fix an issue where some code that was in a cell loop should not have been in a cell loop (the code was not wrong, but slightly less precise than it might have been).