Sacado:  Fix ViewFad conversions for new design.
The new design implementation was allowing conversions from other AD
types to ViewFad, which is not correct.  In addition to being generally
unsafe, this was causing compiler errors in Panzer when the two return
types of the ternary operator differ, with one being ViewFad.

It tooks some time to figure out a way to do this that NVCC accepted.
It seems to not handle inherited constructors disabled through SFINAE
correctly.