Tpetra: explicitly sort entries of add() result, make setAllIndices() expect sorted. (#6405)
* Tpetra: make sure add() result has sorted indices

add() uses setAllValues() to populate the result matrix, but
this does not mark entries as unsorted internally. So the later
fillComplete did not actually sort them. Now, add() sorts entries
before calling setAllValues().

* Tpetra: setAllIndices/setAllValues enforces sorted

CrsGraph::setAllIndices() and CrsGraph::setAllValues() really
assume that the indices being passed in are sorted within rows,
because they do not set the isSorted_ flag to false. This commit
updates the doxygen to reflect this, and (in graph) explicitly
sets isSorted_ and noRedundancies_ to true (because neither sorting
nor merging need to happen during fillComplete).

* Tpetra: use Tpetra::Import_Util::sortCrsEntries

...instead of re-implementing it with Tpetra::sort2