ATDM: Disable packages GEMMA does not use in all cuda+complex builds (ATDV-265)
We don't have the human resources to continue to keep testing and trying to
maintain builds and tests that ATDM does not use.  If ATDM APPs are not
directly impacted, Trilinos developers mostly just ignore the GitHub issues we
create so it is a waste of computer resources and human time maintaining them.

See added notes in this commit for more details.