TSQR::Impl::CuSolver: Take info in constructor, not methods
The goal is for TSQR::Impl::CuSolver and TSQR::Impl::RawQR to have the
same interface.  This will make testing easier.  It might also make it
possible to unify TSQR::CombineNodeTsqr and TSQR::CuSolverNodeTsqr
with a common interface, by having the (still hypothetical) unified
LapackNodeTsqr subclass ask the RawQR implementation whether it wants
device memory for "large" arrays (A, Q, and C).