TSQR: Add class to simplify MPI & Kokkos init & finalize in tests
TSQR::Test::MpiAndKokkosScope is a scope guard for MPI and Kokkos
initialization and finalization.  It works whether or not MPI is
enabled.  (We can't use Tpetra::ScopeGuard because that lives in
TpetraCore, and TpetraCore depends on TpetraTSQR.)  I use the new
class in the DistTsqr test.

I also changed the DistTsqr test in the following ways:

  - it always builds, even if MPI is disabled (this works in TSQR
    via Teuchos::Comm's SerialComm back-end); and
  - it now exercises multiple MPI processes if MPI is enabled.
    (Before, it only ran on 1 MPI process.)