Zoltan2: Fixes for #6440
This will resolve an indeterminancy in the MJ calculation.
Mj tests can now detect indeterminate results and
return an error. Cdash showed a seg fault in one case and
it's not clear if that is related. But this will at least
get the test running in a consistent manner and should clear
most of the random failures taking down auto PR testing.