Tpetra::Details::Behavior: Add longRowMinNumEntries
@trilinos/tpetra

If a local row of a sparse graph or matrix has this many or more
entries, we consider it "long."  Tpetra has the option to treat
"long" rows separately from shorter rows in algorithms like
sparse matrix-vector multiply, packAndPrepare, and
unpackAndCombine.  You may control this at run time via the
TPETRA_LONG_ROW_MIN_NUM_ENTRIES environment variable.

This is an absolute value, not a relative value.  This matters
for thread parallelization, because the point is to decide how
much parallelism needs to be available in a row in order to
treat it differently from other rows.  Whether a row has enough
entries to treat it as "dense" relative to other rows, is a
separate question.