Tpetra,kokkos-kernels: Fix #6633
@trilinos/tpetra

For vectors of `float`, the two-argument overload of KokkosBlas::dot
(that takes two vectors (rank-1 Views) as input, and returns the dot
product result as a scalar value on host) gets inaccurate results on
Mac Clang.  This commit fixes that overload of KokkosBlas::dot so that
if the dot product result type is float resp. complex<float>, the
function uses double resp. complex<double> as the type for
intermediate sums.