Teuchos: Adjust doxygen comments for SWIG 4.0
PyTrilinos is upgrding to SWIG 4.0, which includes the capability to
parse doxygen comments and add them to the Python module as Python
docstrings.  SWIG 4.0 currently has a problem with certain syntax, and
this commit adjusts those doxygen comments to avoid SWIG warnings.
Typical changes are from something like "\f$var\f$" to "\f$ var\f$",
or "\param[in] var description" to "\param var [in] description".